Move KEML io into own service (it could even be a normal static method class, but this might be different in bigger contexts - will analyse, though
